Added interface for TraineeController.
