+++ src/main/java/com/revature/gambit/controllers/TraineeController.java
@@ -2,10 +2,6 @@ package com.revature.gambit.controllers;
 
 import java.util.List;
 
-import org.apache.log4j.Logger;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,24 +9,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 import com.revature.gambit.entities.Trainee;
-import com.revature.gambit.services.TraineeService;
 
-/**
- * Handles all Janus requests for Trainee resources.
- *
- */
-@RestController
-@RequestMapping(value = "trainees", produces = MediaType.APPLICATION_JSON_VALUE)
-public class TraineeController {
-	private static final Logger log = Logger.getLogger(TraineeController.class);
-	
-	@Autowired
-	private TraineeService traineeService;
-	
+public interface TraineeController {
+
 	/**
 	 * Returns all trainees from a batch that has the input batch id and input
 	 * status. Merged two old endpoints into this one. The old endpoint urls were:
@@ -43,23 +26,15 @@ public class TraineeController {
 	 *            - id of the batch desired. status - status of trainees desired.
 	 * @return The list of trainees within that batch with the given batchId.
 	 */
-	@GetMapping("batch/{id}/status/{status}")
+
 	// @PreAuthorize("hasAnyRole('VP', 'QC', 'TRAINER', 'STAGING', 'PANEL')")
-	public ResponseEntity<List<Trainee>> findAllByBatchAndStatus(@PathVariable Integer id,
-			@PathVariable String status) {
-		log.debug("Trainee Controller received request: Finding trainees for batch: " + id + " with status: " + status);
-		return new ResponseEntity<>(traineeService.findAllByBatchAndStatus(id, status), HttpStatus.OK);
-	}
+	ResponseEntity<List<Trainee>> findAllByBatchAndStatus(Integer id, String status);
 
 	/**
 	 * 
 	 * @return a List of all trainees
 	 */
-	@GetMapping
-	public ResponseEntity<List<Trainee>> getAll() {
-		log.debug("Trainee Controller received request: getAll trainees");
-		return new ResponseEntity<List<Trainee>>(traineeService.getAll(), HttpStatus.OK);
-	}
+	ResponseEntity<List<Trainee>> getAll();
 
 	/**
 	 * Creates a new trainee. The old endpoint url was: "/all/trainee/create" In
@@ -69,12 +44,9 @@ public class TraineeController {
 	 *            - the trainee to be created.
 	 * @return The newly created trainee.
 	 */
-	@PostMapping
+
 	// @PreAuthorize("hasAnyRole('VP', 'QC', 'TRAINER', 'PANEL')")
-	public ResponseEntity<Trainee> createTrainee(@RequestBody Trainee trainee) {
-		log.debug("Trainee Controller received request: Creating trainee: " + trainee);
-		return new ResponseEntity<>(traineeService.save(trainee), HttpStatus.CREATED);
-	}
+	ResponseEntity<Trainee> createTrainee(Trainee trainee);
 
 	/**
 	 * Updates the given trainee.
@@ -86,12 +58,9 @@ public class TraineeController {
 	 *            - the trainee to be updated.
 	 * @return A response entity signifying a successful update.
 	 */
-	@PutMapping
+
 	// @PreAuthorize("hasAnyRole('VP', 'QC', 'TRAINER','PANEL')")
-	public ResponseEntity<Trainee> updateTrainee(@RequestBody Trainee trainee) {
-		log.debug("Trainee Controller received request: Updating trainee: " + trainee);
-		return new ResponseEntity<>(traineeService.save(trainee), HttpStatus.NO_CONTENT);
-	}
+	ResponseEntity<Trainee> updateTrainee(Trainee trainee);
 
 	/**
 	 * Deletes the given trainee.
@@ -103,19 +72,11 @@ public class TraineeController {
 	 *            - the id of the trainee to delete
 	 * @return A response entity signifying a successful deletion
 	 */
-	@DeleteMapping("/{traineeId}")
+
 	// @PreAuthorize("hasAnyRole('VP', 'QC', 'TRAINER','PANEL')")
-	public ResponseEntity<Void> deleteTrainee(@PathVariable Integer traineeId) {
-		log.debug("Trainee Controller received request: Deleting trainee: " + traineeId);		
-		traineeService.delete(traineeId);
-		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-	}
-	
-	@GetMapping("/{email}")
+	ResponseEntity<Void> deleteTrainee(Integer traineeId);
+
 	// @PreAuthorize("hasAnyRole('VP', 'QC', 'TRAINER', 'STAGING', 'PANEL')")
-	public ResponseEntity<Trainee> findByEmail(@PathVariable String email) {
-		log.debug("Finding trainees by email: " + email);
-		return new ResponseEntity<>(traineeService.findByEmail(email), HttpStatus.OK);
-	}
+	ResponseEntity<Trainee> findByEmail(String email);
 
 }
